Workaround GNOME/gvfs#249 when using open syscall
Problem: open() may fail with EOPNOTSUPP in linux if opening a file
in a gvfs-fuse mount with O_WRONLY or O_RDWR when neither O_TRUNC nor
O_APPEND is set;
This is a known issue:
[GNOME/gvfs#249](https://gitlab.gnome.org/GNOME/gvfs/-/issues/249)

Solution: Open a file with O_APPEND when not specifically overwriting or
truncating. Memfile functions may still be affected by this bug, sorry.

Closes: https://github.com/vim/vim/issues/5237